pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helixcommit"
version = "0.1.0"
description = "Automated release notes generator for Git repositories."
readme = "README.md"
authors = [
    { name = "HelixCommit" }
]
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "GitPython>=3.1.43",
    "Typer>=0.12.0",
    "Jinja2>=3.1.2",
    "requests>=2.32.3",
    "PyGithub>=2.3.0",
    "openai>=1.52.0",
    "python-dateutil>=2.9.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-mock>=3.14.0",
    "pytest-cov>=5.0.0",
    "responses>=0.25.3",
    "ruff>=0.6.9",
    "mypy>=1.13.0"
]

[project.urls]
Homepage = "https://github.com/bjornefisk/HelixCommit"
Repository = "https://github.com/bjornefisk/HelixCommit"

[project.scripts]
helixcommit = "helixcommit.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py39"
line-length = 100
extend-exclude = ["build", "dist", ".venv"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "B",  # flake8-bugbear
    "I",  # isort
    "RUF",
]
ignore = [
    "E203",  # whitespace before ':'
    "E501",  # allow long strings in prompts/templates
    "B008",  # Typer's option pattern calls at definition time
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false
line-ending = "lf"

[tool.mypy]
python_version = "3.11"
mypy_path = ["src"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra --showlocals --strict-markers --strict-config"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
    "error::DeprecationWarning",
    "error::PendingDeprecationWarning",
]
